WEBtool/requirements.txt
Flask
flask_cors
Pillow
opencv-python